requirements.txt
flask
tensorflow>=2.11
opencv-python
mtcnn
numpy
pillow
werkzeug
scikit-learn
matplotlib
pandas
gunicorn
